Quality improvements

+++ 6-Mathematics/6.6.1_Number_Theoretic_Transform.cpp
@@ -93,7 +93,7 @@ std::vector<int64_t> convolve(std::vector<int64_t> a, std::vector<int64_t> b) {
   if (a.empty() || b.empty()) {
     return {};
   }
-  int result_size = a.size() + b.size() - 1;
+  int result_size = static_cast<int>(a.size() + b.size()) - 1;
   if (std::min(a.size(), b.size()) < NAIVE_CUTOFF) {
     std::vector<int64_t> res(result_size);
     for (int i = 0; i < static_cast<int>(a.size()); i++) {
